Routed UART, buttons, and flipped analog power cap
I needed a capacitor near the power input at the top, because otherwise the power is flowing directly from the plane through the one via. I'm still not sure that I've fixed a problem, its seems likely that the main power flowing so close to the chips pins is a bad idea.