Swapped header logical numbers
I realized my numbering scheme was off with the actual board layout. So in the schematic I swapped the connectors without rewiring so that the numbers are more logical. CS7, CS8, and Free2 are now for the top mezzanine connector, which is closer to the pins on the chips. It also falls in line with my bottom up numbering scheme.