Fixing board edge pin header 3d models
I tried first switching out the footprint entirely, but that resulted in the 3d models pointing in to the board. I rotated all the connections in the schematic and the board, but that came with a billion errors because the pin one is a square and way to many traces are close to the pins. I instead used my current vertical footprint and manually changed the 3d model for each. I may have problems if these headers are put in a pick and place, because they will point inward by default, but in most cases these should be hand soldered.