Redesigned Motor Driver module for higher current
I doubled up the connectors on the edge, and I switched to a cheaper version of the same edge connector from Keystone. I'm planning on cutting the receptor in half because the male connector does not fully use the length, and I can save space and time. This is why the footprint is sticking off the edge